Extract default warehouse logic to hook
